bbsims/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.1
backports.entry-points-selectable==1.1.0
coverage==5.5
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-countries==7.2.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-phone-field==1.8.1
django-polymorphic==3.0.0
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.0.1
platformdirs==2.3.0
psycopg2==2.9.2
PyMySQL==0.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2020.4
reportlab==3.5.68
six==1.15.0
sqlparse==0.4.1
virtualenv==20.1.0
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
whitenoise==5.3.0
wkhtmltopdf==0.2
xhtml2pdf==0.2.5
